Clean up Patch
- Remove clutter from the Patch.scala source file, making it readable for end users.
- Move implementation details to internal package or Patch.internal object.
- Remove v0 methods from Patch object.
- Document remaining Patch methods.
